+++ modules/AWSPowerShell/Cmdlets/CodeArtifact/Basic/Get-CAPackageVersionReadme-Cmdlet.cs
@@ -28,10 +28,7 @@ using Amazon.CodeArtifact.Model;
 namespace Amazon.PowerShell.Cmdlets.CA
 {
     /// <summary>
-    /// Gets the readme file or descriptive text for a package version. For packages that
-    /// do not contain a readme file, CodeArtifact extracts a description from a metadata
-    /// file. For example, from the <code>&lt;description&gt;</code> element in the <code>pom.xml</code>
-    /// file of a Maven package. 
+    /// Gets the readme file or descriptive text for a package version. 
     /// 
     ///  
     /// <para>
@@ -81,7 +78,8 @@ namespace Amazon.PowerShell.Cmdlets.CA
         /// <summary>
         /// <para>
         /// <para> A format that specifies the type of the package version with the requested readme
-        /// file. </para>
+        /// file. </para><note><para>Although <code>maven</code> is listed as a valid value, CodeArtifact does not support
+        /// displaying readme files for Maven packages.</para></note>
         /// </para>
         /// </summary>
         #if !MODULAR