+++ modules/AWSPowerShell/Cmdlets/OpenSearchService/Basic/Add-OSResourceTag-Cmdlet.cs
@@ -29,8 +29,8 @@ namespace Amazon.PowerShell.Cmdlets.OS
 {
     /// <summary>
     /// Attaches tags to an existing Amazon OpenSearch Service domain. Tags are a set of case-sensitive
-    /// key-value pairs. An domain can have up to 10 tags. For more information, see <a href="https://docs.aws.amazon.com/opensearch-service/latest/developerguide/managedomains.html#managedomains-awsresorcetagging">
-    /// Tagging Amazon OpenSearch Service domains</a>.
+    /// key-value pairs. A domain can have up to 10 tags. For more information, see <a href="https://docs.aws.amazon.com/opensearch-service/latest/developerguide/managedomains-awsresourcetagging.html">Tagging
+    /// Amazon OpenSearch Service domains</a>.
     /// </summary>
     [Cmdlet("Add", "OSResourceTag", SupportsShouldProcess = true, ConfirmImpact = ConfirmImpact.Medium)]
     [OutputType("None")]